The FragmentHierarchicalNavigation project was renamed to FragmentNavEx.  This was to remove an error in the build caused by the length of the name.
  Effected files:
    FragmentHierarchicalNavigation.sln
    FragmentHierarchicalNavigation\FragmentHierarchicalNavigation.csproj
    FragmentHierarchicalNavigation\Resources\Resource.Designer.cs
    FragmentHierarchicalNavigation\FragmentNavEx.csproj

The ActionBarSherlock project was added from the Xamarin samples.  Code sharing options were considered from the ones offered in this Xamarin document [1].  Since portable class libraries are not available the ActionBarSherlock code was not included as an solution-external reference.

  Effected Files:
    ActionBarSherlock\ActionBarSherlock.csproj
    ActionBarSherlock\Additions\AboutAdditions.txt
    ActionBarSherlock\Jars\AboutJars.txt
    ActionBarSherlock\Jars\ActionBarSherlock-4.2.0.zip
    ActionBarSherlock\Jars\android-support-v4.jar
    ActionBarSherlock\Properties\AssemblyInfo.cs
    ActionBarSherlock\Transforms\EnumFields.xml
    ActionBarSherlock\Transforms\EnumMethods.xml
    ActionBarSherlock\Transforms\Metadata.xml

The minimum android to target was modified in the ActionBarSherlock to version 2.2.  This was to allow the FragmentNavEx project to be able to reference it.

  Effected Files:
    ActionBarSherlock\ActionBarSherlock.csproj

The ActionBarSherlock Java package was added to the SupportLib folder of the FragmentNavEx project.  This is to allow the build process to place it into the package.

  Effected Files:
    FragmentHierarchicalNavigation\SupportLib\ActionBarSherlock-4.2.0.zip

[1] http://docs.xamarin.com/Android/Guides/Application_Fundamentals/Building_Cross_Platform_Applications/Sharing_Code_Options