The Activity1.cs file was removed, as the Main.cs is now the main launcher. The main launcher sets the content view as the Main layout.
  Effected Files:
    FragmentHierarchicalNavigation\Activity1.cs
    FragmentHierarchicalNavigation\Main.cs

The main layout was changed to include fragments.  The large layout contains two side-by-side fragments, while the regular layout contains only one.

  Effected Files:
    FragmentHierarchicalNavigation\Resources\Layout-large\Main.axml
    FragmentHierarchicalNavigation\Resources\Layout\Main.axml
    FragmentHierarchicalNavigation\Resources\Values\Strings.xml

The fragment classes were added.  FragmentA represents a summary view, while FragmentB represents a detailed view.  The FragmentA view contains a button that opens a FragmentB view.  Both FragmentA and FragmentB extend AbstractFragment, as the two had similar functionality could be abstracted.

  Effected Files:
    FragmentHierarchicalNavigation\AbstractFragment.cs
    FragmentHierarchicalNavigation\FragmentA.cs
    FragmentHierarchicalNavigation\FragmentB.cs

The Fragment layout was added as a common layout for both FragmentA and FragmentB.  It has two buttons which are hidden by FragmentB, and used for navigation by FragmentA.  The TextView is where the output for AbstractFragment.GetOutput is stored.

  Effected Files:
    FragmentHierarchicalNavigation\Resources\Layout\Fragment.axml